requirements.txt
feedparser==6.0.11
beautifulsoup4==4.12.3
python-dotenv==1.0.1

certifi==2024.7.4
